Remove Femiwiki extensions that have migrated to Wikimedia Gerrit
The repos removed in this commit are:
 - https://github.com/femiwiki/AchievementBadges
 - https://github.com/femiwiki/DiscordRCFeed
 - https://github.com/femiwiki/Sanctions

These GitHub repos are either marked as archived, or have a pointer
from their README to their new location on Gerrit. In all cases, the
repository link on their https://mediawiki.org/ extension page has been
updated to point to Gerrit (rather than GitHub).

Now that they've moved to Gerrit, the previous GitHub repos should be
removed from `nonwmf-extensions`; to avoid duplication in systems that
combine the extensions in Gerrit with the submodules in this repo.

Fixes #92